fix(import_columbia): Use courts-db to find courts
Reduce amount of files used
Use match_based_text function from harvard to find duplicated content
Typing added
